+++ NEWS.md
@@ -39,6 +39,8 @@ Version 0.13b2 - dev
 - NEW: `Hatch.paths.ellipse_edges_to_spline_edges()` convert ellipse edges to spline edges
 - NEW: `Hatch.paths.all_to_spline_edges()` convert all curves to approximated spline edges
 - NEW: `Hatch.paths.all_to_line_edges()` convert all curves to approximated line edges
+- NEW: `BSpline.from_arc()` returns a B-Spline from a `ConstructionArc()`
+- NEW: `BSpline.from_ellipse()` returns a B-Spline from a `ConstructionEllipse()`
 - NEW: `BSpline.from_fit_points()` returns a B-Spline from fit points
 - NEW: `BSpline.arc_approximation()` returns a B-Spline approximation for an arc defined by `ConstructionArc()`
 - NEW: `BSpline.ellipse_approximation()` returns a B-Spline approximation for an ellipse defined by `ConstructionEllipse()`